refactor: migrate to plasmo framework
