+++ .gitignore
@@ -1,5 +1,3 @@
 node_modules
-dist
-distribution
-.cache
-.parcel-cache
+build
+.plasmo
