--- .terserrc
@@ -1,6 +1,0 @@
-{
-	"output": {
-		"beautify": true,
-		"indent_level": 2
-	}
-}
