+++ background.ts
@@ -1,0 +1,1 @@
+export { extensionOptions } from '~libs/options-storage';
