+++ options.tsx
@@ -1,0 +1,4 @@
+import 'webext-base-css/webext-base.css';
+import { extensionOptions } from '~libs/options-storage';
+
+extensionOptions.syncForm('form');
