+++ package.json
@@ -1,50 +1,53 @@
 {
-	"private": true,
-	"scripts": {
-		"build": "parcel build source/manifest.json --dist-dir distribution --no-cache --no-source-maps",
-		"lint": "xo",
-		"fix": "xo --fix",
-		"test": "npm run lint && npm run build",
-		"start": "parcel watch source/manifest.json --dist-dir distribution --no-cache --no-hmr"
-	},
-	"browserslist": [
-		"Chrome 115",
-		"Firefox 115"
-	],
-	"xo": {
-		"envs": [
-			"browser",
-			"webextensions"
-		],
-		"extends": "xo-react",
-		"settings": {
-			"react": {
-				"version": "16.13"
-			}
-		},
-		"rules": {
-			"unicorn/prefer-top-level-await": "off"
-		}
-	},
-	"dependencies": {
-		"delegate-it": "^6.1.0",
-		"dom-chef": "^5.1.1",
-		"doma": "^4.0.0",
-		"element-ready": "^7.0.0",
-		"push-form": "^1.0.1",
-		"select-dom": "^9.0.0",
-		"webext-base-css": "^1.4.4",
-		"webext-options-sync": "^4.2.3"
-	},
-	"devDependencies": {
-		"@parcel/config-webextension": "^2.12.1-canary.3330",
-		"eslint-config-xo-react": "^0.27.0",
-		"eslint-plugin-react": "^7.35.0",
-		"eslint-plugin-react-hooks": "^4.6.2",
-		"parcel": "^2.0.0-canary.1705",
-		"xo": "^0.59.3"
-	},
-	"webExt": {
-		"sourceDir": "distribution"
-	}
+    "displayName": "Notifications Preview for GitHub",
+    "version": "24.8.19",
+    "private": true,
+    "packageManager": "pnpm@9.8.0",
+    "scripts": {
+        "dev": "plasmo dev",
+        "build": "plasmo build",
+        "package": "plasmo package",
+        "lint": "xo",
+        "fix": "xo --fix",
+        "test": "npm run lint && npm run build",
+        "start": "parcel watch source/manifest.json --dist-dir distribution --no-cache --no-hmr"
+    },
+    "manifest": {
+        "manifest_version": 3,
+        "permissions": [
+            "storage"
+        ]
+    },
+    "xo": {
+        "envs": [
+            "browser",
+            "webextensions"
+        ],
+        "extends": "xo-react",
+        "settings": {
+            "react": {
+                "version": "16.13"
+            }
+        },
+        "rules": {
+            "unicorn/prefer-top-level-await": "off"
+        }
+    },
+    "dependencies": {
+        "delegate-it": "^6.2.0",
+        "dom-chef": "^5.1.1",
+        "doma": "^4.0.0",
+        "element-ready": "^7.0.0",
+        "plasmo": "0.88.0",
+        "push-form": "^1.0.1",
+        "select-dom": "^9.0.0",
+        "webext-base-css": "^1.4.4",
+        "webext-options-sync": "^4.2.4"
+    },
+    "devDependencies": {
+        "eslint-config-xo-react": "^0.27.0",
+        "eslint-plugin-react-hooks": "^4.6.2",
+        "eslint-plugin-react": "^7.35.0",
+        "xo": "^0.59.3"
+    }
 }
