--- source/libs/utils.js
@@ -1,8 +1,0 @@
-export function empty(element) {
-	element.textContent = '';
-}
-
-// Wait for the timeout, but don't run if tab is not visible
-export function setTimeoutUntilVisible(callback, ms) {
-	return setTimeout(requestAnimationFrame, ms, callback);
-}
