--- source/manifest.json
@@ -1,42 +1,0 @@
-{
-	"name": "Notifications Preview for GitHub",
-	"homepage_url": "https://github.com/tanmayrajani/notifications-preview-github",
-	"version": "0.0.0",
-	"manifest_version": 3,
-	"minimum_chrome_version": "115",
-	"browser_specific_settings": {
-		"gecko": {
-			"id": "notifications-preview-github@tanmayrajani.github.io",
-			"strict_min_version": "115.0"
-		}
-	},
-	"permissions": [
-		"storage"
-	],
-	"description": "Quickly see your notifications in a popup without leaving the current page",
-	"icons": {
-		"128": "icon-128.png"
-	},
-	"content_scripts": [{
-		"matches": [
-			"https://github.com/*"
-		],
-		"css": [
-			"github-notifications-preview.css"
-		],
-		"js": [
-			"github-notifications-preview.jsx"
-		],
-		"run_at": "document_start"
-	}],
-	"background": {
-		"scripts": [
-			"options-storage.js"
-		],
-		"service_worker": "options-storage.js",
-		"type": "module"
-	},
-	"options_ui": {
-		"page": "options.html"
-	}
-}
