--- source/options-storage.js
@@ -1,13 +1,0 @@
-import OptionsSync from 'webext-options-sync';
-
-const optionsStorage = new OptionsSync({
-	defaults: {
-		previewCount: true,
-		markAllAsRead: true,
-		dropdown: 'compact',
-		participating: false,
-		closeOnMouseleave: false,
-	},
-});
-
-export default optionsStorage;
