--- source/options.html
@@ -1,37 +1,0 @@
-<!doctype html>
-<meta charset="utf-8">
-<title>Options • Notifications Preview for GitHub</title>
-<link rel="stylesheet" href="../node_modules/webext-base-css/webext-base.css">
-<form>
-	<p>
-		<label>
-			<input type="checkbox" name="previewCount">
-			Show notification number in blue bubble
-		</label>
-	</p>
-	<p>
-		<label>
-			<input type="checkbox" name="markAllAsRead">
-			Show "Mark all as read" link
-		</label>
-	</p>
-	<p>
-		<label>
-			<input type="checkbox" name="participating">
-			Show only participating notifications
-		</label>
-	</p>
-	<p>
-		<label>
-			<input type="checkbox" name="closeOnMouseleave">
-			Automatically close the popup when mouse is moved away
-		</label>
-	</p>
-	<p>
-		Show notifications list on hover: <br>
-		<label><input type="radio" name="dropdown" value="yes"> Yes</label><br>
-		<label><input type="radio" name="dropdown" value="compact"> Yes, compact (default)</label><br>
-		<label><input type="radio" name="dropdown" value="no"> No</label>
-	</p>
-</form>
-<script src="options.js" type="module"></script>
