--- source/options.js
@@ -1,3 +1,0 @@
-import optionsStorage from './options-storage.js';
-
-optionsStorage.syncForm('form');
